Refactor the driver directory name
