+++ .gitignore
@@ -1,1 +1,1 @@
-strata/cmd/mongo/lreplica_s3storage_driver/strata/strata
+strata/cmd/mongo/lreplica_drivers/strata/strata
