+++ strata/cmd/mongo/lreplica_drivers/strata/main.go
@@ -9,7 +9,7 @@ import (
 	"runtime"
 
 	"github.com/facebookgo/rocks-strata/strata"
-	"github.com/facebookgo/rocks-strata/strata/cmd/mongo/lreplica_s3storage_driver/lrs3driver"
+	"github.com/facebookgo/rocks-strata/strata/cmd/mongo/lreplica_drivers/lrs3driver"
 )
 
 func main() {
